paper-search-be/pyproject.toml
[project]
name = "paper-search-be"
version = "0.1.0"
description = ""
authors = [
    {name = "timechess",email = "tyc221@ruc.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools (>=1.70.0,<2.0.0)",
    "grpcio (>=1.70.0,<2.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "openai (>=1.63.0,<2.0.0)",
    "jsonlines (>=4.0.0,<5.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
